+++ app/src/main/java/hu/bme/aut/fitary/dataSource/UserDAO.kt
@@ -1,0 +1,72 @@
+package hu.bme.aut.fitary.dataSource
+
+import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.database.ChildEventListener
+import com.google.firebase.database.DataSnapshot
+import com.google.firebase.database.DatabaseError
+import com.google.firebase.database.FirebaseDatabase
+import hu.bme.aut.fitary.dataSource.model.UserProfile
+import javax.inject.Singleton
+
+@Singleton
+class UserDAO {
+
+    private val database = FirebaseDatabase.getInstance()
+    private val auth = FirebaseAuth.getInstance()
+
+    val currentUser: UserProfile?
+        get() = users[auth.currentUser?.uid]
+
+    private val _users = mutableMapOf<String, UserProfile>()
+    val users: Map<String, UserProfile>
+        get() = _users.toMap()
+
+    init {
+        database
+            .getReference("users")
+            .addChildEventListener(object : ChildEventListener {
+
+                override fun onChildAdded(dataSnapshot: DataSnapshot, s: String?) {
+                    val newUser = dataSnapshot.getValue(UserProfile::class.java)
+
+                    newUser?.let { _users += Pair(it.id, it) }
+                }
+
+                override fun onChildChanged(
+                    dataSnapshot: DataSnapshot,
+                    previousChildName: String?
+                ) {
+                    val user = dataSnapshot.getValue(UserProfile::class.java)
+                    user?.let { _users.replace(it.id, it) }
+                }
+
+                override fun onChildRemoved(dataSnapshot: DataSnapshot) {
+                    val user = dataSnapshot.getValue(UserProfile::class.java)
+                    user?.let { _users.remove(it.id) }
+                }
+
+                override fun onChildMoved(dataSnapshot: DataSnapshot, previousChildName: String?) {
+                    TODO("Not yet implemented")
+                }
+
+                override fun onCancelled(error: DatabaseError) {
+                    TODO("Not yet implemented")
+                }
+            })
+    }
+
+    suspend fun saveUser(user: UserProfile) {
+        if (_users.containsKey(user.id) && _users[user.id] == user)
+            return
+
+        val key = database.reference
+            .child("users")
+            .push().key ?: return
+
+        database.reference
+            .child("users")
+            .child(key)
+            .setValue(user)
+    }
+
+}
