+++ app/src/main/java/hu/bme/aut/fitary/dataSource/WorkoutDAO.kt
@@ -9,20 +9,15 @@ import com.google.firebase.database.FirebaseDatabase
 import hu.bme.aut.fitary.dataSource.model.Exercise
 import hu.bme.aut.fitary.dataSource.model.UserProfile
 import hu.bme.aut.fitary.dataSource.model.Workout
+import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
-class FirebaseDAO {
+class WorkoutDAO @Inject constructor(
+    private val userDAO: UserDAO
+) {
 
     private val database = FirebaseDatabase.getInstance()
-    private val auth = FirebaseAuth.getInstance()
-
-    val currentUser: UserProfile?
-        get() = users[auth.currentUser?.uid]
-
-    private val _users = mutableMapOf<String, UserProfile>()
-    val users: Map<String, UserProfile>
-        get() = _users.toMap()
 
     private val _workouts = mutableListOf<Workout>()
     val workouts: List<Workout>
@@ -32,15 +27,7 @@ class FirebaseDAO {
     val userWorkouts
         get() = _userWorkouts.toList()
 
-    private val _exercises = mutableMapOf<Long, Exercise>()
-    val exercises: Map<Long, Exercise>
-        get() = _exercises.toMap()
-
     init {
-        initFirebaseConnection()
-    }
-
-    private fun initFirebaseConnection() {
         database
             .getReference("workouts")
             .addChildEventListener(object : ChildEventListener {
@@ -51,7 +38,7 @@ class FirebaseDAO {
                     newWorkout?.let {
                         _workouts += newWorkout
 
-                        if (newWorkout.uid == currentUser?.id)
+                        if (newWorkout.uid == userDAO.currentUser?.id)
                             _userWorkouts += newWorkout
                     }
                 }
@@ -75,38 +62,6 @@ class FirebaseDAO {
                     TODO("Not yet implemented")
                 }
             })
-
-        database
-            .getReference("users")
-            .addChildEventListener(object : ChildEventListener {
-
-                override fun onChildAdded(dataSnapshot: DataSnapshot, s: String?) {
-                    val newUser = dataSnapshot.getValue(UserProfile::class.java)
-
-                    newUser?.let { _users += Pair(it.id, it) }
-                }
-
-                override fun onChildChanged(
-                    dataSnapshot: DataSnapshot,
-                    previousChildName: String?
-                ) {
-                    val user = dataSnapshot.getValue(UserProfile::class.java)
-                    user?.let { _users.replace(it.id, it) }
-                }
-
-                override fun onChildRemoved(dataSnapshot: DataSnapshot) {
-                    val user = dataSnapshot.getValue(UserProfile::class.java)
-                    user?.let { _users.remove(it.id) }
-                }
-
-                override fun onChildMoved(dataSnapshot: DataSnapshot, previousChildName: String?) {
-                    TODO("Not yet implemented")
-                }
-
-                override fun onCancelled(error: DatabaseError) {
-                    TODO("Not yet implemented")
-                }
-            })
     }
 
     suspend fun saveWorkout(workout: Workout) {
@@ -119,21 +74,5 @@ class FirebaseDAO {
             .child(key)
             .setValue(workout)
     }
-
-    suspend fun saveUser(user: UserProfile) {
-        if (_users.containsKey(user.id) && _users[user.id] == user)
-            return
-
-        val key = database.reference
-            .child("users")
-            .push().key ?: return
-
-        database.reference
-            .child("users")
-            .child(key)
-            .setValue(user)
-    }
-
-    suspend fun getExerciseById(id: Long) = exercises[id]
 
 }
